# Final-Project/docs/swagger.yaml
basePath: /
definitions:
...
  MYGRAM-API_user_utils.ResponseMessage:
    properties:
      data:
        example: the error explained here
        type: string
      status:
        example: fail
        type: string
    type: object
...
  utils.LoggedinUser:
    properties:
      token:
        example: the token generated here
        type: string
    type: object
  utils.LoginUser:
    properties:
      email:
        example: saukiadillah@gmail.com
        type: string
      password:
        example: secret
        type: string
    type: object
...
  utils.RegisterUser:
    properties:
      age:
        example: 8
        type: integer
      email:
        example: saukiadillah@gmail.com
        type: string
      password:
        example: secret
        type: string
      username:
        example: sauki
        type: string
    type: object
  utils.RegisteredUser:
    properties:
      age:
        example: 8
        type: integer
      email:
        example: saukiadillah@gmail.com
        type: string
      id:
        example: the user id generated here
        type: string
      username:
        example: sauki
        type: string
    type: object
...
  utils.ResponseDataLoggedinUser:
    properties:
      data:
        $ref: '#/definitions/utils.LoggedinUser'
      status:
        example: success
        type: string
    type: object
  utils.ResponseDataRegisteredUser:
    properties:
      data:
        $ref: '#/definitions/utils.RegisteredUser'
      status:
        example: success
        type: string
    type: object
...
  utils.ResponseDataUpdatedUser:
    properties:
      data:
        $ref: '#/definitions/utils.UpdatedUser'
      status:
        example: success
        type: string
    type: object
...
  utils.ResponseMessageDeletedUser:
    properties:
      message:
        example: your account has been successfully deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  utils.UpdateUser:
    properties:
      email:
        example: newsaukiadillah@gmail.com
        type: string
      username:
        example: sauki
        type: string
    type: object
...
  utils.UpdatedUser:
    properties:
      age:
        example: 8
        type: integer
      email:
        example: newsaukiadillah@gmail.com
        type: string
      id:
        example: here is the generated user id
        type: string
      updated_at:
        example: the updated at generated here
        type: string
      username:
        example: sauki
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: saukiadillah@gmail.com
    name: sauki
  description: MyGram is a free photo sharing app written in Go. People can share,
    view, and comment photos by everyone. Anyone can create an account by registering
    an email address and creating a username.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: MyGram API
  version: "1.0"
paths:
...
  /users:
    delete:
      consumes:
      - application/json
      description: Delete a user with authentication user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseMessageDeletedUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/MYGRAM-API_user_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/MYGRAM-API_user_utils.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/MYGRAM-API_user_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Delete a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user with authentication user
      parameters:
      - description: Update User
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseDataUpdatedUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/MYGRAM-API_user_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/MYGRAM-API_user_utils.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/MYGRAM-API_user_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Update a user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authentication a user and retrieve a token
      parameters:
      - description: Login User
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseDataLoggedinUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/MYGRAM-API_user_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/MYGRAM-API_user_utils.ResponseMessage'
      summary: Login a user
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: create and store a user
      parameters:
      - description: Register User
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.RegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ResponseDataRegisteredUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/MYGRAM-API_user_utils.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/MYGRAM-API_user_utils.ResponseMessage'
      summary: Register a user
      tags:
      - users
securityDefinitions:
  Bearer:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"